+++ src/components/recruit/FaqSection/FaqSection.tsx
