+++ src/components/recruit/FaqSection/index.tsx
